requirements.txt
apify-client>=1.3.0
httpx>=0.27.0
yt-dlp>=2023.12.30